# reduce the amount of healing