--- a/internal/config/scanner/scanner.go
+++ b/internal/config/scanner/scanner.go
@@ -99,7 +99,7 @@ func LookupConfig(kvs config.KVS) (cfg Config, err error) {
 	case "fast":
 		cfg.Delay, cfg.MaxWait, cfg.Cycle = 1, 100*time.Millisecond, time.Minute
 	case "default":
-		cfg.Delay, cfg.MaxWait, cfg.Cycle = 2, 5*time.Second, time.Minute
+		cfg.Delay, cfg.MaxWait, cfg.Cycle = 2, time.Second, time.Minute
 	case "slow":
 		cfg.Delay, cfg.MaxWait, cfg.Cycle = 10, 15*time.Second, time.Minute
 	case "slowest":